media: ov5645: Improve mode finding function
Find the sensor mode by comparing the size of the requested image size
and the sensor mode's image size. The distance between image sizes is the
size in pixels of the non-overlapping regions between the requested size
and the frame-specified size. This logic is borrowed from et8ek8 sensor
driver.

Signed-off-by: Todor Tomov <todor.tomov@linaro.org>